vararg spread optimization working with handling for hoisting
