+++ src/lualib/ArraySplice.ts
@@ -1,5 +1,5 @@
 // https://www.ecma-international.org/ecma-262/9.0/index.html#sec-array.prototype.splice
-function __TS__ArraySplice<T>(this: void, list: T[], ...args: Vararg<unknown[]>): T[] {
+function __TS__ArraySplice<T>(this: void, list: T[], ...args: unknown[]): T[] {
     const len = list.length;
 
     const actualArgumentCount = select("#", ...args);
