+++ src/lualib/Generator.ts
@@ -8,7 +8,7 @@ function __TS__GeneratorIterator(this: GeneratorIterator) {
     return this;
 }
 
-function __TS__GeneratorNext(this: GeneratorIterator, ...args: Vararg<any[]>) {
+function __TS__GeneratorNext(this: GeneratorIterator, ...args: any[]) {
     const co = this.____coroutine;
     if (coroutine.status(co) === "dead") return { done: true };
 
@@ -19,7 +19,7 @@ function __TS__GeneratorNext(this: GeneratorIterator, ...args: Vararg<any[]>) {
 }
 
 function __TS__Generator(this: void, fn: (this: void, ...args: any[]) => any) {
-    return function (this: void, ...args: Vararg<any[]>): GeneratorIterator {
+    return function (this: void, ...args: any[]): GeneratorIterator {
         const argsLength = select("#", ...args);
         return {
             // Using explicit this there, since we don't pass arguments after the first nil and context is likely to be nil
