+++ src/lualib/New.ts
@@ -1,4 +1,4 @@
-function __TS__New(this: void, target: LuaClass, ...args: Vararg<any[]>): any {
+function __TS__New(this: void, target: LuaClass, ...args: any[]): any {
     const instance: any = setmetatable({}, target.prototype);
     instance.____constructor(...args);
     return instance;
